+++ backend/canisters/user/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support getting batches of chat events via LocalUserIndex ([#4848](https://github.com/open-chat-labs/open-chat/pull/4848))
+- Add `c2c_send_messages_v2` which uses MessageContentInternal ([#4902](https://github.com/open-chat-labs/open-chat/pull/4902))
 
 ### Changed